add ability to run on random data for analysis purposes

+++ src/BruteForceSolver.java
@@ -16,7 +16,7 @@ class BruteForceSolver {
 
     private int numberOfInitialPoints;
 
-    private int distanceCalculations = 0;
+    private long distanceCalculations = 0;
 
     /**
      * Constructs a BruteForceSolver object that runs the brute force
